Initial implementation of an OSX sandbox.
The basic idea is that we create a sandbox directory, and hard link all the
spawn inputs into that directory. Then we run the command there, and finally
copy the results back.

This CL is somewhat of a hack, as it changes the linux sandboxing code in
place rather than making a copy. This needs to be cleaned up.

When the basic implementation is in place, we can further lock it down with
OSX's sandboxing mechanism, sandbox-exec. Right now, access to absolute
paths is unrestricted, but at least the temporary exec root makes it harder
to accidentally access other files in the workspace.